api/requirements.txt
fastapi>=0.115.12
uvicorn[standard]>=0.34.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.50
PyYAML>=6.0.1
pytz>=2024.1
